### 2023-02-14 #5151 @editorialbot
[PRE REVIEW]: SolvationAnalysis: A Python toolkit for understanding liquid solvation structure in classical molecular dynamics simulations
**Submitting author:** @orionarcher (<a href="http://orcid.org/0000-0003-3940-2456">Orion Cohen</a>)
**Repository:** https://github.com/MDAnalysis/solvation-analysis
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
