+++ .github/workflows/CI.yaml
@@ -77,7 +77,7 @@ jobs:
         name: codecov-${{ matrix.os }}-py${{ matrix.python-version }}
   
   test_pypi:
-    name: Test on ${{ matrix.os }}, Python ${{ matrix.python-version }}
+    name: pip (PEP517) install on ${{ matrix.os }}, Python ${{ matrix.python-version }}
     runs-on: ${{ matrix.os }}
     strategy:
       fail-fast: false
