+++ .github/workflows/CI.yaml
@@ -76,7 +76,7 @@ jobs:
         flags: unittests
         name: codecov-${{ matrix.os }}-py${{ matrix.python-version }}
   
-  test_pypi:
+  test_pip_install:
     name: pip (PEP517) install on ${{ matrix.os }}, Python ${{ matrix.python-version }}
     runs-on: ${{ matrix.os }}
     strategy:
