+++ setup.py
@@ -45,7 +45,7 @@ setup(
     setup_requires=[] + pytest_runner,
 
     install_requires=[
-        'numpy>=1.16.0',
+        'numpy>=1.20.0',
         'mdanalysis>=2.0.0b0',
         'pandas',
         'matplotlib',
