### 2023-02-14 9821dc4
add rdkit req and fix pip
