+++ setup.py
@@ -51,7 +51,8 @@ setup(
         'matplotlib',
         'scipy',
         'statsmodels', 
-        'plotly'
+        'plotly',
+        'rdkit'
     ],
     # Additional entries you may want simply uncomment the lines you want and fill in the data
     # url='http://www.my_package.com',  # Website
