+++ requirements.txt
@@ -8,3 +8,4 @@ matplotlib
 scipy
 statsmodels
 plotly
+rdkit
