+++ .github/workflows/CI.yaml
@@ -88,12 +88,17 @@ jobs:
     steps:
     - uses: actions/checkout@v3
 
-    - name: Install package
+    - uses: conda-incubator/setup-miniconda@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+        activate-environment: test
 
+    - name: Install package
+      shell: bash -l {0}
       run: |
         python -m pip install .
 
     - name: Run tests
-
+      shell: bash -l {0}
       run: |
         pip install pytest && pytest -v --color=yes solvation_analysis/tests/
