### 2023-02-15 2710e04
change MDA pin and switch to setup-python
