+++ devtools/conda-envs/test_env.yaml
@@ -10,7 +10,7 @@ dependencies:
     # Pip-only installs
   - pip:
       # new MDAnalysis
-      - MDAnalysis==2.3.0
+      - MDAnalysis>=2.0.0
       # base sci-python libraries
       - scipy
       - numpy
