+++ devtools/conda-envs/test_env.yaml
@@ -3,23 +3,16 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-    # Base depends
-  - python
+  - python>=3.8
   - pip
-
-    # Pip-only installs
-  - pip:
-      # new MDAnalysis
-      - MDAnalysis>=2.0.0
-      # base sci-python libraries
-      - scipy
-      - numpy
-      - pandas
-      - matplotlib
-      - statsmodels
-      # testing #
-      - pytest
-      - pytest-cov
-      - codecov
-      - rdkit
-      - plotly
+  - MDAnalysis>=2.0.0
+  - scipy
+  - numpy
+  - pandas
+  - matplotlib
+  - statsmodels
+  - pytest
+  - pytest-cov
+  - codecov
+  - rdkit
+  - plotly
