### 2023-02-15 9461f84
Add pyproject.toml (#86)
* minimal pyproject.toml

* add versioneer[toml]

* update CI

* change name

* Irfan suggestions

* rename run

* finish pyproj including deps

* fix typo

* fix typo

* enclose in quotes

* fix install_requires

* fix numpy pin

* add rdkit req and fix pip

* add blank conda env

* change MDA pin and switch to setup-python

* fix pin and move to conda install

* remove pytest install
