+++ setup.py
@@ -45,11 +45,14 @@ setup(
     setup_requires=[] + pytest_runner,
 
     install_requires=[
-        'numpy>=1.16.0',
+        'numpy>=1.20.0',
         'mdanalysis>=2.0.0b0',
         'pandas',
         'matplotlib',
         'scipy',
+        'statsmodels', 
+        'plotly',
+        'rdkit'
     ],
     # Additional entries you may want simply uncomment the lines you want and fill in the data
     # url='http://www.my_package.com',  # Website
