### 2023-03-07 #5183 @arose
@orionarcher et al., I find this is a very useful, well documented piece of software. A few comments

- Please emphasize more that the tutorials need to taken from the repository to work interactively. I first tried it with my own notebook, but eventually found that there are file dependencies in the repositories. Opening the notebooks from the repository worked like a charm.
- I installed the package with `pip install solvation_analysis` but that version did not had rdkit and plotly as dependencies which already has been fixed in the repository. Please note in the project readme how to best install the package.
- I did not find community guidelines as outlines in the review checklist. Maybe it is self-evident being a github project but an xtra sentence in the readme does not hurt.
- A note about nglview would be nice since it is used for the figure and has a chapter in the tutorial (disclosure, I am an author of nglview)
