### 2023-03-07 f8c1af4
update installation settings
