+++ setup.py
@@ -46,11 +46,14 @@ setup(
 
     install_requires=[
         'numpy>=1.16.0',
-        'mdanalysis>=2.0.0b0',
+        'mdanalysis>=2.0.0',
         'pandas',
         'matplotlib',
         'scipy',
-    ],
+        'statsmodels',
+        'plotly',
+        'rdkit',
+        ],
     # Additional entries you may want simply uncomment the lines you want and fill in the data
     # url='http://www.my_package.com',  # Website
     # install_requires=[],              # Required packages, pulls from pip if needed; do not use for Conda deployment
