### 2023-03-07 6c5dc7b
added contributing section to README
