+++ joss_paper/paper.bib
@@ -102,6 +102,23 @@ doi          = {10.5281/zenodo.3509134},
 url          = {https://doi.org/10.5281/zenodo.3509134},
 }
 
+@article{nglview:2018,
+    author = {Nguyen, Hai and Case, David A and Rose, Alexander S},
+    title = "{NGLview–interactive molecular graphics for Jupyter notebooks}",
+    journal = {Bioinformatics},
+    volume = {34},
+    number = {7},
+    pages = {1241-1242},
+    year = {2017},
+    month = {12},
+    issn = {1367-4803},
+    doi = {10.1093/bioinformatics/btx789},
+    url = {https://doi.org/10.1093/bioinformatics/btx789},
+    eprint = {https://academic.oup.com/bioinformatics/article-pdf/34/7/1241/48914829/bioinformatics\_34\_7\_1241.pdf},
+}
+
+
+
 @article{Hou:2019,
 title = {The influence of FEC on the solvation structure and reduction reaction of LiPF6/EC electrolytes and its implication for solid electrolyte interphase formation},
 author = {Hou, Tingzheng and Yang, Guang and Rajput, Nav Nidhi and Self, Julian and Park, Sang-Won and Nanda, Jagjit and Persson, Kristin A.},
