### 2023-03-07 #5183 @orionarcher
Hi @arose, thanks so much for the timely review. I've incorporated your feedback into a PR [here](https://github.com/MDAnalysis/solvation-analysis/pull/87). Points 2-4 are clear and have been addressed.

For point 1, I removed several hard links in `clustering_and_residence_tutorial.ipynb` (good catch), everything now imports from `datafiles.py`, which isn't hard coded. I believe the tutorials should now run from anywhere and not just inside the repo. Does that fix the issue for you or does it still need to be run from the local notebook copy?

Happy to push the new version to PyPI or make additional changes first, whichever is best. This is my first submission to JOSS so I'm not sure what the best workflow is.
