### 2023-03-14 #5183 @arose
> I've incorporated your feedback into a PR https://github.com/MDAnalysis/solvation-analysis/pull/87. Points 2-4 are clear and have been addressed.

Thank you, looks good to me.

> For point 1, I removed several hard links in clustering_and_residence_tutorial.ipynb (good catch), everything now imports from datafiles.py, which isn't hard coded. I believe the tutorials should now run from anywhere and not just inside the repo. Does that fix the issue for you or does it still need to be run from the local notebook copy?

That is great, thank you. All good for me.

> Happy to push the new version to PyPI or make additional changes first, whichever is best. This is my first submission to JOSS so I'm not sure what the best workflow is.

As long as there is a new version before the paper is published I am happy.
