### 2023-03-29 #5183 @amritagos
I think this is a nice, fairly well-documented package that would further the ultimate goal of reproducible research. Many of the analyses are very useful, and I especially like the visualization and plotting features. I was able to install the package using `pip` and run the examples on my local machine and cluster. However, as previously mentioned, it might be nice to have some more documentation about how to install the package.

I am happy to recomment this for publication.
