### 2023-03-30 #5183 @zhubonan
Thanks for @arose and @amritagos for reviewing the package.

@orionarcher Could you please implement the suggestions from the reviewers. I can see the PR 👏here: https://github.com/MDAnalysis/solvation-analysis/pull/87, but it is not merged yet.

Please note that we will require a new tagged release before acceptance. The release should be archived in a repository with a DOI. The easiest way to do this is through [Zenodo](https://zenodo.org/). Once you have done this please post the new version and the DOI here.
