### 2023-04-02 3f784a9
turn coordination propertys to @property methods
