+++ solvation_analysis/tests/test_solute.py
@@ -175,7 +175,7 @@ def test_speciation_find_shells(shell, n_shells, run_solute):
 )
 def test_coordination_numbers(name, cn, run_solute):
     # duplicated to test in solute
-    coord_dict = run_solute.coordination.cn_dict
+    coord_dict = run_solute.coordination.coordination_numbers
     np.testing.assert_allclose(cn, coord_dict[name], atol=0.05)
 
 
