### 2023-04-02 537b9da
update networking with @property methods
