### 2023-04-02 ec78b4f
changed pairing attributes to @property
