+++ solvation_analysis/tests/test_pairing.py
@@ -7,7 +7,7 @@ from solvation_analysis.pairing import Pairing
 
 def test_pairing_from_solute(run_solute):
     pairing = Pairing.from_solute(run_solute)
-    assert len(pairing.pairing_dict) == 3
+    assert len(pairing.solvent_pairing) == 3
     assert len(pairing.fraction_free_solvents) == 3
 
 
@@ -21,7 +21,7 @@ def test_pairing_from_solute(run_solute):
 )
 def test_pairing_dict(name, fraction, solvation_data):
     pairing = Pairing(solvation_data, 10, 49, {'fec': 237, 'bn': 363, 'pf6': 49})
-    np.testing.assert_allclose(fraction, pairing.pairing_dict[name], atol=0.05)
+    np.testing.assert_allclose(fraction, pairing.solvent_pairing[name], atol=0.05)
     assert len(pairing.pairing_by_frame) == 3
 
 
@@ -48,6 +48,6 @@ def test_pairing_participating(name, fraction, solvation_data):
 )
 def test_diluent_composition(name, diluent_fraction, solvation_data):
     pairing = Pairing(solvation_data, 10, 49, {'fec': 237, 'bn': 363, 'pf6': 49})
-    np.testing.assert_allclose(diluent_fraction, pairing.diluent_dict[name], atol=0.05)
-    np.testing.assert_allclose(sum(pairing.diluent_dict.values()), 1, atol=0.05)
+    np.testing.assert_allclose(diluent_fraction, pairing.diluent_composition[name], atol=0.05)
+    np.testing.assert_allclose(sum(pairing.diluent_composition.values()), 1, atol=0.05)
 
