### 2023-04-02 ba67bdd
move @property methods to end of class
