### 2023-04-02 02c904d
update properties in residence
