### 2023-04-02 58f70e8
delete attribute doc strings
