+++ solvation_analysis/tests/test_speciation.py
@@ -6,7 +6,7 @@ from solvation_analysis.speciation import Speciation
 
 def test_speciation_from_solute(run_solute):
     speciation = Speciation.from_solute(run_solute)
-    assert len(speciation.speciation_data) == 490
+    assert len(speciation.speciation_df) == 490
 
 
 @pytest.mark.parametrize(
@@ -49,7 +49,7 @@ def test_speciation_find_shells(shell, n_shells, solvation_data):
 )
 def test_speciation_correlation(solvent_one, solvent_two, correlation, solvation_data):
     speciation = Speciation(solvation_data, 10, 49)
-    df = speciation.co_occurrence
+    df = speciation.solvent_co_occurrence
     np.testing.assert_allclose(df[solvent_one][solvent_two], correlation, atol=0.05)
 
 
