### 2023-04-02 9b96772
reorder residence methods
