+++ solvation_analysis/residence.py
@@ -83,6 +83,15 @@ class Residence:
         {'BN': 4.02, 'FEC': 3.79, 'PF6': 1.15}
     """
 
+    def __init__(self, solvation_data, step):
+        self.solvation_data = solvation_data
+        self._auto_covariances = self._calculate_auto_covariance_dict()
+        self._residence_times_cutoff = self._calculate_residence_times_with_cutoff(self._auto_covariances, step)
+        self._residence_times_fit, self._fit_parameters = self._calculate_residence_times_with_fit(
+            self._auto_covariances,
+            step
+        )
+
     @staticmethod
     def from_solute(solute):
         """
@@ -270,12 +279,3 @@ class Residence:
         arameters for the exponential fit to the autocorrelation function.
         """
         return self._fit_parameters
-
-    def __init__(self, solvation_data, step):
-        self.solvation_data = solvation_data
-        self._auto_covariances = self._calculate_auto_covariance_dict()
-        self._residence_times_cutoff = self._calculate_residence_times_with_cutoff(self._auto_covariances, step)
-        self._residence_times_fit, self._fit_parameters = self._calculate_residence_times_with_fit(
-            self._auto_covariances,
-            step
-        )
