### 2023-04-02 #5183 @orionarcher
Thanks for the review @amritagos!

@zhubonan I merged the PR which incorporated all of the requested changes.

I just pushed `v0.3.1` to PyPI and conda-forge. The zenodo reference is (here)[https://zenodo.org/record/7793127#.ZCmwtezMKdY] and has DOI: 10.5281/zenodo.7793127
