+++ solvation_analysis/tests/test_speciation.py
@@ -52,9 +52,3 @@ def test_speciation_correlation(solvent_one, solvent_two, correlation, solvation
     df = speciation.solvent_co_occurrence
     np.testing.assert_allclose(df[solvent_one][solvent_two], correlation, atol=0.05)
 
-
-def test_plot_correlation(solvation_data):
-    speciation = Speciation(solvation_data, 10, 49)
-    fig, ax = speciation.plot_co_occurrence()
-    # fig.show()
-
