### 2023-04-06 b721d96
update plotting tutorial
