### 2023-04-06 #5183 @orionarcher
Done! `v0.3.2` is available on PyPI and Github.
