### 2023-04-15 #5183 @Kevin-Mattheus-Moerman
@zhubonan FYI you can now call `@editorialbot create post-review checklist` at the end of review to generate this list ☝️
