### 2023-04-17 #5183 @zhubonan
> @zhubonan FYI you can now call `@editorialbot create post-review checklist` at the end of review to generate this list ☝️ You can also see I've left some unticked, e.g. the license on the ZENODO archive does not match. I'll ask the authors to fix this now.

Got it, I will do that the next time! Thanks for the help!
